requirements.txt
flask
requests
cryptography
flask_sqlalchemy
pillow
python-dotenv
google-generativeai